+++ api/main.go
@@ -14,14 +14,15 @@ func main() {
 	var config = &Config{}
 	flag.StringVar(&config.LogLevel, "LogLevel", "debug", "logging threshold level: debug|info|warn|error|fatal|panic")
 	flag.IntVar(&config.Port, "Port", 8888, "HTTP port to listen on")
-	flag.StringVar(&config.RethinkDB, "RethinkDB", "localhost:28015", "RethinkDB server addresses comma delimited")
-	flag.StringVar(&config.Database, "Database", "syros", "RethinkDB database name")
+	flag.StringVar(&config.MongoDB, "MongoDB", "localhost:27017", "MongoDB server addresses comma delimited")
+	flag.StringVar(&config.Database, "Database", "syros", "MongoDB database name")
 	flag.StringVar(&config.JwtSecret, "JwtSecret", "syros", "JWT secret")
 	flag.StringVar(&config.Credentials, "Credentials", "admin@admin", "Credentials format user@password")
 	flag.StringVar(&config.AppPath, "AppPath", "", "Path to dist dir")
 	flag.Parse()
 
 	setLogLevel(config.LogLevel)
+
 	log.Infof("Starting with config: %+v", config)
 
 	if config.AppPath == "" {
@@ -40,7 +41,7 @@ func main() {
 
 	repo, err := NewRepository(config)
 	if err != nil {
-		log.Fatalf("RethinkDB connection error %v", err)
+		log.Fatalf("MongoDB connection error %v", err)
 	}
 
 	server := HttpServer{