+++ indexer/main.go
@@ -18,10 +18,9 @@ func main() {
 	flag.StringVar(&config.CollectorQueue, "CollectorQueue", "syros", "Nats collector queue name")
 	flag.StringVar(&config.RegistryTopic, "RegistryTopic", "registry", "Nats registry topic name")
 	flag.StringVar(&config.RegistryQueue, "RegistryQueue", "syros", "Nats registry queue name")
-	flag.StringVar(&config.RethinkDB, "RethinkDB", "localhost:28015", "RethinkDB server addresses comma delimited")
-	flag.StringVar(&config.Database, "Database", "syros", "RethinkDB database name")
+	flag.StringVar(&config.MongoDB, "MongoDB", "localhost:27017", "MongoDB server addresses comma delimited")
+	flag.StringVar(&config.Database, "Database", "syros", "MongoDB database name")
 	flag.IntVar(&config.DatabaseStale, "DatabaseStale", 5, "Deletes database records older than specified value in minutes, set 0 to disable")
-	flag.IntVar(&config.DatabaseStaleSince, "DatabaseStaleSince", 48, "Scan for database records since specified value in hours")
 	flag.Parse()
 
 	setLogLevel(config.LogLevel)
@@ -29,17 +28,13 @@ func main() {
 
 	repo, err := NewRepository(config)
 	if err != nil {
-		log.Fatalf("RethinkDB connection error %v", err)
+		log.Fatalf("MongoDB connection error %v", err)
 	}
 
 	repo.Initialize()
-	log.Infof("Connected to RethinkDB cluster %v database initialization done", config.RethinkDB)
+	log.Infof("Connected to MongoDB cluster %v database initialization done", config.MongoDB)
 
-	gcrepo, err := NewRepository(config)
-	if err != nil {
-		log.Fatalf("RethinkDB connection error %v", err)
-	}
-	gcrepo.RunGarbageCollector([]string{"containers", "hosts", "checks", "syros_services"})
+	repo.RunGarbageCollector([]string{"containers", "hosts", "checks", "syros_services"})
 
 	nc, err := NewNatsConnection(config.Nats)
 	if err != nil {
@@ -49,11 +44,7 @@ func main() {
 
 	log.Infof("Connected to NATS server %v status %v", nc.ConnectedUrl(), nc.Status())
 
-	regrepo, err := NewRepository(config)
-	if err != nil {
-		log.Fatalf("RethinkDB connection error %v", err)
-	}
-	registry := NewRegistry(config, nc, regrepo)
+	registry := NewRegistry(config, nc, repo)
 	registry.WatchForAgents()
 
 	indexer := models.SyrosService{
@@ -79,9 +70,7 @@ func main() {
 	consumer.Consume()
 
 	server := &HttpServer{
-		Config:     config,
-		Registry:   registry,
-		Repository: repo,
+		Config: config,
 	}
 	log.Infof("Starting HTTP server on port %v", config.Port)
 	go server.Start()